Extract ItemsNotInFolder from project folder settings
