+++ front/app/modules/commercial/project_folders/admin/containers/projects/index.tsx
@@ -1,14 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { isNilOrError } from 'utils/helperUtils';
 import { withRouter, WithRouterProps } from 'react-router';
 
 // services
-import { PublicationStatus } from 'services/projects';
-import { updateProjectFolderMembership } from '../../../services/projects';
 import { isAdmin } from 'services/permissions/roles';
 
 // hooks
-import useAdminPublications from 'hooks/useAdminPublications';
 import useAuthUser from 'hooks/useAuthUser';
 
 // localisation
@@ -16,10 +13,9 @@ import { FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
 
 // components
-import { List, Row } from 'components/admin/ResourceList';
 import { HeaderTitle } from 'containers/Admin/projects/all/StyledComponents';
-import ProjectRow from 'containers/Admin/projects/components/ProjectRow';
 import ItemsInFolder from './ItemsInFolder';
+import ItemsNotInFolder from './ItemsNotInFolder';
 
 // style
 import styled from 'styled-components';
@@ -53,42 +49,16 @@ const Spacer = styled.div`
   flex: 1;
 `;
 
-const publicationStatuses: PublicationStatus[] = [
-  'draft',
-  'archived',
-  'published',
-];
-
 const AdminFolderProjectsList = ({
   params: { projectFolderId },
 }: WithRouterProps) => {
-  const { list: allPublications } = useAdminPublications({
-    publicationStatusFilter: publicationStatuses,
-  });
-
   const authUser = useAuthUser();
-  const [processing, setProcessing] = useState<string[]>([]);
 
   if (isNilOrError(authUser)) {
     return null;
   }
 
-  const addProjectToFolder = (projectFolderId: string) => (
-    projectId: string
-  ) => async () => {
-    setProcessing([...processing, projectId]);
-    await updateProjectFolderMembership(projectId, projectFolderId);
-    setProcessing(processing.filter((item) => projectId !== item));
-  };
-
   const userIsAdmin = authUser && isAdmin({ data: authUser });
-
-  const otherProjects = !isNilOrError(allPublications)
-    ? allPublications.filter(
-        (item) =>
-          item.publicationType === 'project' && item.attributes.depth === 0
-      )
-    : null;
 
   return (
     <Container>
@@ -110,37 +80,7 @@ const AdminFolderProjectsList = ({
               </StyledHeaderTitle>
             </ListHeader>
 
-            {otherProjects ? (
-              <List>
-                <>
-                  {otherProjects.map((adminPublication, index: number) => (
-                    <Row
-                      id={adminPublication.id}
-                      isLastItem={index === otherProjects.length - 1}
-                      key={adminPublication.id}
-                    >
-                      <ProjectRow
-                        publication={adminPublication}
-                        actions={[
-                          {
-                            buttonContent: (
-                              <FormattedMessage {...messages.addToFolder} />
-                            ),
-                            handler: addProjectToFolder(projectFolderId),
-                            processing: processing.includes(
-                              adminPublication.publicationId
-                            ),
-                            icon: 'plus-circle',
-                          },
-                        ]}
-                      />
-                    </Row>
-                  ))}
-                </>
-              </List>
-            ) : (
-              <FormattedMessage {...messages.noProjectsToAdd} />
-            )}
+            <ItemsNotInFolder projectFolderId={projectFolderId} />
           </>
         )}
       </ListsContainer>
